refactor: rename dubbo_resolver package to dubbo
