+++ pkg/filter/http/remote/resolver/dubbo/dubbo_resolver_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package dubbo_resolver
+package dubbo
 
 import (
 	"context"
